+++ juliaset/__init__.py
@@ -1,3 +1,4 @@
 from .juliaset import julia
+from .juliaset import JuliaSet
 
 __version__ = "0.3.0-dev"